Filter by Appointments Total Cost and some other changes
1. Added an another filter option for Total Payment of a customer. But,
facing a problem in Model. I Calculated all appointments Cost for each
Customer (SUM(total_cost) AS total_amount). this "total_amount" is
working in view page but is not working in Model Page. That's why
getting an error. Is there any solution for it?
Please help!

2. Staff Crud has been solved.

there are some other changes too, please download them.